requirements.txt
appnope==0.1.0
decorator==4.1.2
Django==1.11.6
django-extensions==1.9.1
djangorestframework==3.7.0
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
jsonfield==2.0.2
mysqlclient==1.3.7
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2017.2
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
